+++ jmh-benchmarks/src/main/java/org/apache/kafka/jmh/metadata/MetadataRequestBenchmark.java
@@ -174,7 +174,7 @@ public class MetadataRequestBenchmark {
         kafkaProps.put(KafkaConfig$.MODULE$.BrokerIdProp(), brokerId + "");
         KafkaConfig config = new KafkaConfig(kafkaProps);
         return new KafkaApis(requestChannel,
-            new ZkSupport(adminManager, kafkaController, kafkaZkClient, Option.empty(), metadataCache, config),
+            new ZkSupport(adminManager, kafkaController, kafkaZkClient, Option.empty(), metadataCache),
             replicaManager,
             groupCoordinator,
             transactionCoordinator,
